Organize file structure for components
