+++ src/Components/RecipeCard/RecipeCard.js
@@ -1,0 +1,1 @@
+import './RecipeCard.css';
