+++ src/Components/RecipeDetails/RecipeDetails.js
@@ -1,0 +1,1 @@
+import './RecipeDetail.css';
